+++ irs_rrt/rrt_params.py
@@ -1,4 +1,6 @@
+import numpy as np
 from irs_mpc.irs_mpc_params import BundleMode
+
 
 
 class RrtParams:
@@ -42,3 +44,16 @@ class IrsRrtParams(RrtParams):
         # Stepsize.
         # TODO(terry-suh): the selection of this parameter should be automated.
         self.stepsize = 0.3
+
+class IrsRrtGlobalParams(IrsRrtParams):
+    def __init__(self, q_model_path, joint_limits):
+        super().__init__(q_model_path, joint_limits)
+        # Global distance metric for defining an adequate notion of distance.
+        # Should be a vector of size dim_x.a
+        self.global_metric = np.isnan
+
+class IrsRrtRolloutParams(IrsRrtGlobalParams):
+    def __init__(self, q_model_path, joint_limits):
+        super().__init__(q_model_path, joint_limits)
+        # Rollout horizon to use for reaching the point with a long stepsize.
+        self.rollout_horizon = 3
